Improve fn_800A0AB4 95.4->96.3% (drop (int) cast on f32->u8 store)

+++ src/main/dll/dll_000B_dll0b.c
@@ -2583,7 +2583,7 @@ animate:
         int ofs = k * 4 + 0xb0;
         for (j = 0; j < ((ModgfxVertexGroupCmd*)p)->indexCount; j++)
         {
-            bufB[(*(s16**)((char*)p + 0x10))[j] * 16 + 0xf] = (int)*(f32*)((char*)state + ofs);
+            bufB[(*(s16**)((char*)p + 0x10))[j] * 16 + 0xf] = *(f32*)((char*)state + ofs);
             bufA[(*(s16**)((char*)p + 0x10))[j] * 16 + 0xf] =
                 bufB[(*(s16**)((char*)p + 0x10))[j] * 16 + 0xf];
         }
